streaming version of GetShopifyIds, with time and total count parameters.
